+++ packages/entities/entities-routes/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.19.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.19.1...@kong-ui-public/entities-routes@3.19.2) (2026-02-12)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.19.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.19.0...@kong-ui-public/entities-routes@3.19.1) (2026-02-11)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
 # [3.19.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.18.8...@kong-ui-public/entities-routes@3.19.0) (2026-02-09)